track_by_detection/requirements.txt
numpy
numba
opencv-python
matplotlib
scipy
tqdm
multiprocessing
joblib
tensorflow-gpu==2.x
"git+https://github.com/cvxgrp/cvxpy.git#egg=cvxpy[glpk]"
cvxopt